agent: add upload logs
